cleanup ME test folder
